+++ test/unit/fuzz-tools.test.js
@@ -417,6 +417,7 @@ import * as goutallierV437 from '../../lib/goutallier-v437.js';
 import * as eatonLittlerV438 from '../../lib/eaton-littler-v438.js';
 import * as hamadaV439 from '../../lib/hamada-v439.js';
 import * as barrowCcfV440 from '../../lib/barrow-ccf-v440.js';
+import * as bordenDavfV441 from '../../lib/borden-davf-v441.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -779,6 +780,7 @@ const MODULES = {
   'eaton-littler-v438.js': eatonLittlerV438,
   'hamada-v439.js': hamadaV439,
   'barrow-ccf-v440.js': barrowCcfV440,
+  'borden-davf-v441.js': bordenDavfV441,
   'tb-testing.js': tbTesting,
 };